refactoring product price

+++ src/main/java/com/magazinejornada/api/controller/request/ProductRequest.java
@@ -1,9 +1,12 @@
 package com.magazinejornada.api.controller.request;
 
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -14,7 +17,7 @@ public class ProductRequest {
     private String title;
     private String description;
     private Long quantity;
-    private Double price;
+    private BigDecimal price;
     private String picture;
 
 }
